Alpha Cronback scan

the function provide the variation of relaibility of alpha when an item is drop from the data.

Usage

alpha.scan(data, items, digits = 3, parallel = FALSE, verbose = TRUE)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

digits

Number of digits in the table summary

parallel

Bool use doSNOW and parallel when there is a great number of items

verbose

Bool for progression visualisation

Value

return a table with the variation of the alpha when the item is remove from the pool

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
alpha.scan(inference, items)
#> progres: 1/8
#> progres: 2/8
#> progres: 3/8
#> progres: 4/8
#> progres: 5/8
#> progres: 6/8
#> progres: 7/8
#> progres: 8/8
#>      item   alpha.gain          z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.15958740 0.12310842  0.7129858         0.6950819
#> 2 item_04  0.021728220 0.90593128 0.18248611  0.7193171         0.7015661
#> 3 item_05  0.037101998 1.47240387 0.07045592  0.7039434         0.6852800
#> 4 item_06  0.051736003 1.98905631 0.02334749  0.6893094         0.6683304
#> 5 item_07  0.037809488 1.43522502 0.07561149  0.7032359         0.6823281
#> 6 item_08 -0.002314538 0.09882932 0.46063689  0.7433599         0.7262425
#> 7 item_09  0.022516602 0.91620917 0.17977860  0.7185288         0.7011074
#> 8 item_10  0.018467493 0.73914881 0.22990832  0.7225779         0.7037360
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7295968 0.7410454       0.72326     0.7557875
#> 2         0.7355038 0.7410454       0.72326     0.7557875
#> 3         0.7224433 0.7410454       0.72326     0.7557875
#> 4         0.7075841 0.7410454       0.72326     0.7557875
#> 5         0.7224273 0.7410454       0.72326     0.7557875
#> 6         0.7586294 0.7410454       0.72326     0.7557875
#> 7         0.7366334 0.7410454       0.72326     0.7557875
#> 8         0.7403415 0.7410454       0.72326     0.7557875
alpha.scan(inference, items, parallel = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%
#>      item   alpha.gain         z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.1236245 0.13058618  0.7129858         0.6944074
#> 2 item_04  0.021728220 0.8811291 0.18912398  0.7193171         0.7011395
#> 3 item_05  0.037101998 1.4855870 0.06869420  0.7039434         0.6857432
#> 4 item_06  0.051736003 1.9916950 0.02320227  0.6893094         0.6678951
#> 5 item_07  0.037809488 1.4703042 0.07073969  0.7032359         0.6832556
#> 6 item_08 -0.002314538 0.1013548 0.45963442  0.7433599         0.7280280
#> 7 item_09  0.022516602 0.9058944 0.18249589  0.7185288         0.6989506
#> 8 item_10  0.018467493 0.7335702 0.23160537  0.7225779         0.7028819
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7308377 0.7410454     0.7244256     0.7571125
#> 2         0.7367418 0.7410454     0.7244256     0.7571125
#> 3         0.7221796 0.7410454     0.7244256     0.7571125
#> 4         0.7069281 0.7410454     0.7244256     0.7571125
#> 5         0.7216204 0.7410454     0.7244256     0.7571125
#> 6         0.7586028 0.7410454     0.7244256     0.7571125
#> 7         0.7350731 0.7410454     0.7244256     0.7571125
#> 8         0.7398436 0.7410454     0.7244256     0.7571125