Factor comparison

Compute anova and fit index to compare multiple factor analyse fit for multiple factors solution

factor_comparison(data, items, factors)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items

factors

a facto class object from factor_explorer() or a vector with the number of factor

Value

a table with a CHI² comparison and CFI, TLI, RMSEA and SRMR differences

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
f <- factor_explorer(inference, items)
factor_comparison(inference, items, f)
#> Loading required namespace: GPArotation
#> Warning: convergence not obtained in GPFoblq. 1000 iterations used.
#> Model 1 = psych::fa(r = data[, items], nfactors = factor)
#> Model 2 = psych::fa(r = data[, items], nfactors = factor)
#> Model 3 = psych::fa(r = data[, items], nfactors = factor)
#> 
#> 
#>  df   d.df   chiSq   d.chiSq     PR   test   empirical   d.empirical   test.echi       BIC   d.BIC     CFI     TLI   RMSEA    SRMR   d.CFI   d.TLI   d.RMSEA   d.SRMR
#> ---  -----  ------  --------  -----  -----  ----------  ------------  ----------  --------  ------  ------  ------  ------  ------  ------  ------  --------  -------
#>  20     NA   53.12        NA     NA     NA       60.13            NA          NA   -101.65      NA   0.988   0.983   0.027   0.022      NA      NA        NA       NA
#>  13      7   13.46     39.66   0.00   5.67       14.92         45.21        6.46    -87.14   14.51   1.000   1.000   0.004   0.011   0.012   0.016    -0.023   -0.011
#>   2     11    0.29     13.16   0.28   1.20        0.27         14.65        1.33    -15.18   71.96   1.001   1.009   0.000   0.001   0.001   0.009    -0.004   -0.009
factor_comparison(inference, items, c(1,2,3))
#> Model 1 = psych::fa(r = data[, items], nfactors = factor)
#> Model 2 = psych::fa(r = data[, items], nfactors = factor)
#> Model 3 = psych::fa(r = data[, items], nfactors = factor)
#> 
#> 
#>  df   d.df   chiSq   d.chiSq     PR   test   empirical   d.empirical   test.echi       BIC   d.BIC     CFI     TLI   RMSEA    SRMR   d.CFI   d.TLI   d.RMSEA   d.SRMR
#> ---  -----  ------  --------  -----  -----  ----------  ------------  ----------  --------  ------  ------  ------  ------  ------  ------  ------  --------  -------
#>  20     NA   53.12        NA     NA     NA       60.13            NA          NA   -101.65      NA   0.988   0.983   0.027   0.022      NA      NA        NA       NA
#>  13      7   13.46     39.66   0.00   5.67       14.92         45.21        6.46    -87.14   14.51   1.000   1.000   0.004   0.011   0.012   0.016    -0.023   -0.011
#>   7      6    3.33     10.13   0.12   1.69        3.11         11.81        1.97    -50.84   36.30   1.001   1.005   0.000   0.005   0.001   0.006    -0.004   -0.006