Factor explorer

the function use parallel analyse, Velicer and VSS method to provide the possible number of factor.

factor_explorer(data, items, plot = FALSE, ...)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items

plot

Plot vss and parallel

...

Value

the result provide a print of the factor detected for each of the method

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
factor_explorer(inference, items)
#> $`paralle analysis`
#> Call: psych::fa.parallel(x = data[, items], fa = "fa", plot = plot)
#> Parallel analysis suggests that the number of factors =  2  and the number of components =  NA 
#> 
#>  Eigen Values of 
#> 
#>  eigen values of factors
#> [1]  2.21  0.12  0.04  0.02 -0.01 -0.04 -0.05 -0.08
#> 
#>  eigen values of simulated factors
#> [1]  0.40  0.06  0.04  0.02  0.00 -0.02 -0.04 -0.06
#> 
#>  eigen values of components 
#> [1] 2.91 0.90 0.86 0.74 0.72 0.66 0.65 0.57
#> 
#>  eigen values of simulated components
#> [1] NA
#> 
#> $`vss analysis`
#> 
#> Very Simple Structure
#> Call: psych::vss(x = data[, items], plot = plot)
#> VSS complexity 1 achieves a maximimum of 0.65  with  1  factors
#> VSS complexity 2 achieves a maximimum of 0.72  with  4  factors
#> 
#> The Velicer MAP achieves a minimum of 0.02  with  1  factors 
#> BIC achieves a minimum of  -101.65  with  1  factors
#> Sample Size adjusted BIC achieves a minimum of  -45.84  with  2  factors
#> 
#> Statistics by number of factors 
#>   vss1 vss2   map dof   chisq    prob sqresid  fit  RMSEA  BIC SABIC complex
#> 1 0.65 0.00 0.022  20 5.3e+01 7.8e-05     4.3 0.65 0.0269 -102 -38.1     1.0
#> 2 0.41 0.67 0.059  13 1.3e+01 4.1e-01     4.0 0.67 0.0039  -87 -45.8     1.7
#> 3 0.45 0.71 0.104   7 3.3e+00 8.5e-01     3.2 0.74 0.0000  -51 -28.6     1.6
#> 4 0.45 0.72 0.172   2 2.9e-01 8.6e-01     3.1 0.75 0.0000  -15  -8.8     1.7
#> 5 0.37 0.58 0.313  -2 2.0e-06      NA     3.6 0.71     NA   NA    NA     2.3
#> 6 0.37 0.59 0.462  -5 4.1e-08      NA     3.6 0.71     NA   NA    NA     2.3
#> 7 0.37 0.58 1.000  -7 0.0e+00      NA     3.5 0.71     NA   NA    NA     2.3
#> 8 0.37 0.58    NA  -8 0.0e+00      NA     3.5 0.71     NA   NA    NA     2.3
#>    eChisq    SRMR  eCRMS eBIC
#> 1 6.0e+01 2.2e-02 0.0256  -95
#> 2 1.5e+01 1.1e-02 0.0158  -86
#> 3 3.1e+00 4.9e-03 0.0098  -51
#> 4 2.7e-01 1.4e-03 0.0054  -15
#> 5 1.8e-06 3.8e-06     NA   NA
#> 6 3.6e-08 5.3e-07     NA   NA
#> 7 6.9e-13 2.3e-09     NA   NA
#> 8 6.9e-13 2.3e-09     NA   NA
#> 
#> $factor
#> $factor$parallele
#> [1] 2
#> 
#> $factor$`VSS complexity 1`
#> [1] 1
#> 
#> $factor$`VSS complexity 2`
#> [1] 4
#> 
#> $factor$`Velicer MAP`
#> [1] 1
#> 
#> $factor$BIC
#> [1] 1
#> 
#> $factor$eBIC
#> [1] 1
#> 
#> 
#> attr(,"class")
#> [1] "facto"