Classical Test Theory Summary

the fonction provide a warpping of the freq_scan, discrimination_cor, alpha.scan functions # nolint: line_length_linter.

Usage

CTT_summary(data, items, digits = 3, verbose = TRUE, ...)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

digits

number of digits for result

...

Value

a summary of the result and a list with the complet result of each function

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
CTT_summary(inference, items)
#> frequence computation 
#> item discrimination (rir,rit) computation
#> alpha computation 
#> progres: 1/8
#> progres: 2/8
#> progres: 3/8
#> progres: 4/8
#> progres: 5/8
#> progres: 6/8
#> progres: 7/8
#> progres: 8/8
#>  Results summary : 
#> 
#> item        freq       r   somers.d   polyserial   alpha.gain   p.value   drop.alpha
#> --------  ------  ------  ---------  -----------  -----------  --------  -----------
#> item_03    0.766   0.445      0.566        0.615        0.028     0.127        0.713
#> item_04    0.815   0.413      0.564        0.600        0.022     0.188        0.719
#> item_05    0.713   0.490      0.588        0.650        0.037     0.076        0.704
#> item_06    0.708   0.563      0.684        0.746        0.052     0.028        0.689
#> item_07    0.542   0.489      0.560        0.615        0.038     0.073        0.703
#> item_08    0.596   0.291      0.314        0.369       -0.002     0.461        0.743
#> item_09    0.503   0.417      0.472        0.523        0.023     0.181        0.719
#> item_10    0.318   0.393      0.497        0.513        0.018     0.238        0.723