Alpha Cronback scan

the function provide the variation of relaibility of alpha when an item is drop from the data.

alpha.scan(data, items, digits = 3, parallel = FALSE, verbose = TRUE)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

digits

Number of digits in the table summary

parallel

Bool use doSNOW and parallel when there is a great number of items

verbose

Bool for progression visualisation

Value

return a table with the variation of the alpha when the item is remove from the pool

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
alpha.scan(inference, items)
#> progres: 1/8
#> progres: 2/8
#> progres: 3/8
#> progres: 4/8
#> progres: 5/8
#> progres: 6/8
#> progres: 7/8
#> progres: 8/8
#>      item   alpha.gain          z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.08813382 0.13826802  0.7129858         0.6931927
#> 2 item_04  0.021728220 0.85416026 0.19650810  0.7193171         0.7002006
#> 3 item_05  0.037101998 1.41267719 0.07887533  0.7039434         0.6844717
#> 4 item_06  0.051736003 1.93075755 0.02675652  0.6893094         0.6675774
#> 5 item_07  0.037809488 1.45146745 0.07332487  0.7032359         0.6837341
#> 6 item_08 -0.002314538 0.09499846 0.46215802  0.7433599         0.7258069
#> 7 item_09  0.022516602 0.86565164 0.19334061  0.7185288         0.6994760
#> 8 item_10  0.018467493 0.72615364 0.23387229  0.7225779         0.7032592
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7304646 0.7410454     0.7232362     0.7573716
#> 2         0.7365403 0.7410454     0.7232362     0.7573716
#> 3         0.7230013 0.7410454     0.7232362     0.7573716
#> 4         0.7074896 0.7410454     0.7232362     0.7573716
#> 5         0.7217003 0.7410454     0.7232362     0.7573716
#> 6         0.7591997 0.7410454     0.7232362     0.7573716
#> 7         0.7373420 0.7410454     0.7232362     0.7573716
#> 8         0.7395823 0.7410454     0.7232362     0.7573716
alpha.scan(inference, items, parallel = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%
#>      item   alpha.gain          z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.13119597 0.12898631  0.7129858         0.6941431
#> 2 item_04  0.021728220 0.88672699 0.18761295  0.7193171         0.7015990
#> 3 item_05  0.037101998 1.44977501 0.07356064  0.7039434         0.6847480
#> 4 item_06  0.051736003 1.94813769 0.02569924  0.6893094         0.6692158
#> 5 item_07  0.037809488 1.44145255 0.07472844  0.7032359         0.6825642
#> 6 item_08 -0.002314538 0.09729901 0.46124447  0.7433599         0.7268620
#> 7 item_09  0.022516602 0.88397167 0.18835575  0.7185288         0.6991837
#> 8 item_10  0.018467493 0.70484309 0.24045395  0.7225779         0.7011062
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7294837 0.7410454     0.7241033     0.7574901
#> 2         0.7361226 0.7410454     0.7241033     0.7574901
#> 3         0.7221806 0.7410454     0.7241033     0.7574901
#> 4         0.7091474 0.7410454     0.7241033     0.7574901
#> 5         0.7216579 0.7410454     0.7241033     0.7574901
#> 6         0.7594052 0.7410454     0.7241033     0.7574901
#> 7         0.7363019 0.7410454     0.7241033     0.7574901
#> 8         0.7401245 0.7410454     0.7241033     0.7574901