Alpha Cronback scan

the function provide the variation of relaibility of alpha when an item is drop from the data.

alpha.scan(data, items, digits = 3, parallel = FALSE, verbose = TRUE)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

digits

Number of digits in the table summary

parallel

Bool use doSNOW and parallel when there is a great number of items

verbose

Bool for progression visualisation

Value

return a table with the variation of the alpha when the item is remove from the pool

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
alpha.scan(inference, items)
#> progres: 1/8
#> progres: 2/8
#> progres: 3/8
#> progres: 4/8
#> progres: 5/8
#> progres: 6/8
#> progres: 7/8
#> progres: 8/8
#>      item   alpha.gain         z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.1037146 0.13485848  0.7129858         0.6930987
#> 2 item_04  0.021728220 0.8949276 0.18541289  0.7193171         0.7017420
#> 3 item_05  0.037101998 1.5340411 0.06250977  0.7039434         0.6855322
#> 4 item_06  0.051736003 1.9307658 0.02675601  0.6893094         0.6666951
#> 5 item_07  0.037809488 1.5087247 0.06568457  0.7032359         0.6839705
#> 6 item_08 -0.002314538 0.1014831 0.45958349  0.7433599         0.7272494
#> 7 item_09  0.022516602 0.9045939 0.18284027  0.7185288         0.6997125
#> 8 item_10  0.018467493 0.7337542 0.23154927  0.7225779         0.7027594
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7313556 0.7410454     0.7253208      0.757246
#> 2         0.7370303 0.7410454     0.7253208      0.757246
#> 3         0.7205730 0.7410454     0.7253208      0.757246
#> 4         0.7083959 0.7410454     0.7253208      0.757246
#> 5         0.7212980 0.7410454     0.7253208      0.757246
#> 6         0.7585380 0.7410454     0.7253208      0.757246
#> 7         0.7366026 0.7410454     0.7253208      0.757246
#> 8         0.7403648 0.7410454     0.7253208      0.757246
alpha.scan(inference, items, parallel = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%
#>      item   alpha.gain         z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.1486550 0.12534913  0.7129858         0.6950080
#> 2 item_04  0.021728220 0.8973186 0.18477446  0.7193171         0.7008049
#> 3 item_05  0.037101998 1.4754463 0.07004630  0.7039434         0.6842633
#> 4 item_06  0.051736003 2.0432089 0.02051588  0.6893094         0.6689434
#> 5 item_07  0.037809488 1.5074377 0.06584925  0.7032359         0.6820481
#> 6 item_08 -0.002314538 0.1024547 0.45919789  0.7433599         0.7273917
#> 7 item_09  0.022516602 0.9220330 0.17825569  0.7185288         0.6987630
#> 8 item_10  0.018467493 0.7392150 0.22988824  0.7225779         0.7030244
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7313889 0.7410454     0.7256075     0.7567324
#> 2         0.7366332 0.7410454     0.7256075     0.7567324
#> 3         0.7224776 0.7410454     0.7256075     0.7567324
#> 4         0.7075982 0.7410454     0.7256075     0.7567324
#> 5         0.7200997 0.7410454     0.7256075     0.7567324
#> 6         0.7588832 0.7410454     0.7256075     0.7567324
#> 7         0.7351244 0.7410454     0.7256075     0.7567324
#> 8         0.7408239 0.7410454     0.7256075     0.7567324