Alpha Cronback scan

the function provide the variation of relaibility of alpha when an item is drop from the data.

Usage

alpha.scan(data, items, digits = 3, parallel = FALSE, verbose = TRUE)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

digits

Number of digits in the table summary

parallel

Bool use doSNOW and parallel when there is a great number of items

verbose

Bool for progression visualisation

Value

return a table with the variation of the alpha when the item is remove from the pool

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
alpha.scan(inference, items)
#> progres: 1/8
#> progres: 2/8
#> progres: 3/8
#> progres: 4/8
#> progres: 5/8
#> progres: 6/8
#> progres: 7/8
#> progres: 8/8
#>      item   alpha.gain          z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.12720848 0.12982717  0.7129858         0.6945729
#> 2 item_04  0.021728220 0.85929860 0.19508790  0.7193171         0.7011327
#> 3 item_05  0.037101998 1.42655214 0.07685451  0.7039434         0.6845029
#> 4 item_06  0.051736003 1.96459729 0.02473043  0.6893094         0.6694945
#> 5 item_07  0.037809488 1.48862787 0.06829269  0.7032359         0.6837535
#> 6 item_08 -0.002314538 0.09706563 0.46133714  0.7433599         0.7268858
#> 7 item_09  0.022516602 0.89590617 0.18515144  0.7185288         0.6998107
#> 8 item_10  0.018467493 0.71994791 0.23577854  0.7225779         0.7028708
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7302497 0.7410454      0.723738      0.757018
#> 2         0.7378560 0.7410454      0.723738      0.757018
#> 3         0.7231150 0.7410454      0.723738      0.757018
#> 4         0.7089463 0.7410454      0.723738      0.757018
#> 5         0.7207748 0.7410454      0.723738      0.757018
#> 6         0.7596982 0.7410454      0.723738      0.757018
#> 7         0.7361277 0.7410454      0.723738      0.757018
#> 8         0.7405544 0.7410454      0.723738      0.757018
alpha.scan(inference, items, parallel = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%
#>      item   alpha.gain          z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.08833786 0.13822299  0.7129858         0.6932629
#> 2 item_04  0.021728220 0.87875154 0.18976800  0.7193171         0.7014859
#> 3 item_05  0.037101998 1.47152631 0.07057442  0.7039434         0.6852031
#> 4 item_06  0.051736003 2.00413294 0.02252791  0.6893094         0.6698043
#> 5 item_07  0.037809488 1.46337184 0.07168283  0.7032359         0.6838440
#> 6 item_08 -0.002314538 0.09870373 0.46068675  0.7433599         0.7271653
#> 7 item_09  0.022516602 0.89054906 0.18658557  0.7185288         0.6986516
#> 8 item_10  0.018467493 0.70056208 0.24178818  0.7225779         0.7013798
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7310137 0.7410454     0.7230925     0.7566833
#> 2         0.7364184 0.7410454     0.7230925     0.7566833
#> 3         0.7214482 0.7410454     0.7230925     0.7566833
#> 4         0.7076407 0.7410454     0.7230925     0.7566833
#> 5         0.7217396 0.7410454     0.7230925     0.7566833
#> 6         0.7585336 0.7410454     0.7230925     0.7566833
#> 7         0.7350854 0.7410454     0.7230925     0.7566833
#> 8         0.7406365 0.7410454     0.7230925     0.7566833