Alpha Cronback scan

the function provide the variation of relaibility of alpha when an item is drop from the data.

Usage

alpha.scan(data, items, digits = 3, parallel = FALSE, verbose = TRUE)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

digits

Number of digits in the table summary

parallel

Bool use doSNOW and parallel when there is a great number of items

verbose

Bool for progression visualisation

Value

return a table with the variation of the alpha when the item is remove from the pool

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
alpha.scan(inference, items)
#> progres: 1/8
#> progres: 2/8
#> progres: 3/8
#> progres: 4/8
#> progres: 5/8
#> progres: 6/8
#> progres: 7/8
#> progres: 8/8
#>      item   alpha.gain          z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.11481245 0.13246540  0.7129858         0.6947509
#> 2 item_04  0.021728220 0.85640596 0.19588663  0.7193171         0.7004373
#> 3 item_05  0.037101998 1.43071598 0.07625582  0.7039434         0.6843955
#> 4 item_06  0.051736003 1.94549844 0.02585751  0.6893094         0.6668228
#> 5 item_07  0.037809488 1.44556102 0.07415018  0.7032359         0.6826340
#> 6 item_08 -0.002314538 0.09612547 0.46171046  0.7433599         0.7264358
#> 7 item_09  0.022516602 0.88145944 0.18903460  0.7185288         0.6996422
#> 8 item_10  0.018467493 0.70904414 0.23914854  0.7225779         0.7038997
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7308271 0.7410454     0.7246782     0.7583254
#> 2         0.7370522 0.7410454     0.7246782     0.7583254
#> 3         0.7224903 0.7410454     0.7246782     0.7583254
#> 4         0.7066277 0.7410454     0.7246782     0.7583254
#> 5         0.7213103 0.7410454     0.7246782     0.7583254
#> 6         0.7595267 0.7410454     0.7246782     0.7583254
#> 7         0.7367169 0.7410454     0.7246782     0.7583254
#> 8         0.7422900 0.7410454     0.7246782     0.7583254
alpha.scan(inference, items, parallel = TRUE)
#>   |                                                                              |                                                                      |   0%  |                                                                              |=                                                                     |   1%  |                                                                              |=                                                                     |   2%  |                                                                              |==                                                                    |   3%  |                                                                              |===                                                                   |   4%  |                                                                              |====                                                                  |   5%  |                                                                              |====                                                                  |   6%  |                                                                              |=====                                                                 |   7%  |                                                                              |======                                                                |   8%
#>      item   alpha.gain         z    p.value drop.alpha drop.alpha.lwr.ci
#> 1 item_03  0.028059522 1.1115389 0.13316822  0.7129858         0.6931677
#> 2 item_04  0.021728220 0.8639234 0.19381498  0.7193171         0.6999125
#> 3 item_05  0.037101998 1.4385186 0.07514348  0.7039434         0.6835368
#> 4 item_06  0.051736003 1.9603411 0.02497797  0.6893094         0.6683857
#> 5 item_07  0.037809488 1.4173688 0.07818756  0.7032359         0.6823873
#> 6 item_08 -0.002314538 0.1004122 0.46000856  0.7433599         0.7279488
#> 7 item_09  0.022516602 0.9131066 0.18059323  0.7185288         0.7004545
#> 8 item_10  0.018467493 0.7191820 0.23601439  0.7225779         0.7037746
#>   drop.alpha.upr.ci    alpha0 alpha0.lwr.ci alpha0.upr.ci
#> 1         0.7296612 0.7410454     0.7239387     0.7573482
#> 2         0.7361580 0.7410454     0.7239387     0.7573482
#> 3         0.7214737 0.7410454     0.7239387     0.7573482
#> 4         0.7078748 0.7410454     0.7239387     0.7573482
#> 5         0.7226040 0.7410454     0.7239387     0.7573482
#> 6         0.7583600 0.7410454     0.7239387     0.7573482
#> 7         0.7353791 0.7410454     0.7239387     0.7573482
#> 8         0.7414150 0.7410454     0.7239387     0.7573482