Discrimination correlation

The function provide correlation item rest (rir) or correlation item test (rit) for a all of the item in the test. Multiple type of item corrleation are provided as somer's d, point biserial correlation and polyserial correlation for non dichotomic item Result could be interpretend as the item ability to discriminate between a lower and higher difficulty for this item in regard to the overall raw score.

discrimination_cor(
  data,
  items,
  type = "rir",
  method = c("cor", "somer", "polyserial"),
  conf.level = 0.95,
  ...
)

Arguments

data

a data.frame containing the test or scale response in columns

items

The name of the items (should be >1 )

type

type correlation could be rir for Rest Item-correlation or rit for Total Item-correlation (can be both)

method

the type of correlation to use c("cor","somer","polyserial")

conf.level

the confidence interval for the correlation needed

Value

a data.frame with the result and confidence interval

Examples

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
discrimination_cor(inference, items)
#>   type    item         r  r.lwr.ci  r.upr.ci  somers.d somers.lwr.ci
#> 1  rir item_03 0.4450732 0.4116532 0.4772977 0.5659784     0.5015010
#> 2  rir item_04 0.4125828 0.3780483 0.4459706 0.5635974     0.4818141
#> 3  rir item_05 0.4896669 0.4579252 0.5201617 0.5878897     0.5347608
#> 4  rir item_06 0.5634606 0.5348756 0.5907573 0.6841167     0.6327604
#> 5  rir item_07 0.4893683 0.4576148 0.5198752 0.5601048     0.5226987
#> 6  rir item_08 0.2912004 0.2533020 0.3282064 0.3143698     0.2683658
#> 7  rir item_09 0.4171450 0.3827616 0.4503745 0.4723415     0.4327335
#> 8  rir item_10 0.3927835 0.3576144 0.4268401 0.4971675     0.4517085
#>   somers.upr.ci polyserial
#> 1     0.6304557  0.6148742
#> 2     0.6453806  0.6000359
#> 3     0.6410186  0.6502255
#> 4     0.7354729  0.7462634
#> 5     0.5975109  0.6146025
#> 6     0.3603738  0.3688786
#> 7     0.5119494  0.5228201
#> 8     0.5426265  0.5128318